01/12/2023
Version 1.1.2

Now the name of the app is DLQuick (Not DLQuicker, because DLQuick is quicker)

At the start, it prompts you with a choice of tasks. The final content of result.txt depends on such choice.
To do this, I changed ReferenceFinder in order to make it gather and hold all the DLQ info in separate refContainers. All info except the comparison, which is performed and held by the Analyser.

Good night, then! (or rather good morning...)